Test navigation of all site categories
